requirements.txt
telethon>=1.34.0
aiosqlite>=0.19.0
requests>=2.31.0
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.14.0

